refactor: Equation Explorer graph representation
In anticipation of a follow-up change that allows equation explorer to
look at either the general graph or the finite graph, I have refactored
the equation explorer to use a separate graph JSON for it's graph
representation. This makes it easier to swap out the graph
representation later.

This change should have no functional changes to the equation explorer,
it is just a refactor.
